src/user_management/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
psycopg2-binary
djangorestframework-simplejwt
